+++ changes/bug-7222_run-background-zmq-subscriber
@@ -1,0 +1,1 @@
+- Run the zmq log subscriber in the background to avoid hitting the zmq's buffer limits. Closes #7222.
